Free page management (#33)
* free page management

* shaping memory page

* it builds but does not work

* slowly making sense out of COW over MemoryPage

* pushing towards a single level registry of abandoned pages

* some extractions and tests

* fix tests for now

* format

* remove notion of minimal from the base batch

* step by step, moving closer to the reuse

* abandoned initial implementation done

* abandoned initial implementation done with a failing test

* proper offsets

* more tests

* simplification of the algo

* simplifying

* towards green tests

* fixed!

* docs

+++ src/Paprika/Pages/Page.cs
@@ -21,13 +21,13 @@ public interface IPage
 [StructLayout(LayoutKind.Explicit, Pack = 1, Size = Size)]
 public struct PageHeader
 {
-    public const int Size = sizeof(long);
+    public const int Size = sizeof(uint);
 
     /// <summary>
     /// The id of the last transaction that wrote to this page.
     /// </summary>
     [FieldOffset(0)]
-    public long BatchId;
+    public uint BatchId;
 }
